+++ CHANGES.rst
@@ -1,9 +1,48 @@
-=================
-Ratest changelogs
-=================
+==============
+Latest changes
+==============
+
+next
+====
+
+:date: future
+
+Features
+--------
+
+* Split sections from document for search (useful for refs of page).
+* Add flag: search from database instead of Sphinx bundled search feature.
+
+v0.0.3
+======
+
+:date: 2022-11-26
+
+Update documents
+
+Documentation
+-------------
+
+* Add badges into readme and documentation.
+* Follow notes for v0.0.2
+
+v0.0.2
+======
+
+:date: 2022-11-23
+
+Supporting for local development.
+
+Miscellaneous
+-------------
+
+* Use ``bump2version`` for release.
+* Configure about line-length for ``pycodestyle``.
 
 v0.0.1
 ======
+
+:date: 2022-11-23
 
 This is first release.